structuring initial game components
