+++ components/game.js
@@ -1,0 +1,11 @@
+import React, { Component } from 'react';
+import { View } from 'react-native';
+
+export default class Game extends Component {
+	render() {
+		return (
+			<View>
+			</View>
+		)
+	}
+}
